rearranged tests and added test runner
